feat/PRI-84 Layer division of the back end.
